LayoutTests/imported/w3c/web-platform-tests/css/css-lists/parsing/list-style-computed.sub-expected.txt


FAIL Property list-style value 'outside none none' assert_equals: expected "none" but got "outside none none"
FAIL Property list-style value 'inside none disc' assert_equals: expected "inside" but got "inside none disc"
FAIL Property list-style value 'outside url("https://localhost/") disc' assert_equals: expected "url(\"https://localhost/\")" but got "outside url(\"https://localhost/\") disc"
FAIL Property list-style value 'outside none square' assert_equals: expected "square" but got "outside none square"
PASS Property list-style value 'inside url("https://localhost/") square'
